Implement User Login API

+++ backend/src/main/java/com/Revature/RevStay/daos/UserRepository.java
@@ -2,9 +2,12 @@ package com.Revature.RevStay.daos;
 
 import com.Revature.RevStay.models.User;
 import org.springframework.data.jpa.repository.JpaRepository;
+
 import org.springframework.stereotype.Repository;
 
+import java.util.Optional;
+
 @Repository
-public interface UserRepository extends JpaRepository<User, Long> {
+public interface UserRepository extends JpaRepository<User, Integer> {
     User findByEmail(String email);
 }
